+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.24.0] 2021-11-05
+
+### Added
+
+- `routeToEvents()` function for simple conversion from interaction definitions to messaging events
+- `setState` array operators now supports `,` as separator of items (can be excaped by `\` backslach)
+
 ## [3.23.1] 2021-11-02
 
 ### Fixed